pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python_tooling_enterpy_2021"
version = "0.1.0"
description = "A simple demo project"
authors = ["maces <noreply>"]
readme = "README.md"
keywords = []
packages = [ { include = "python_tooling_enterpy_2021" } ]
include = [
    "README.md",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = "^3.9.4"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
python-decouple = "^3.4"
python-dotenv = "0.15.0"

[tool.poetry.dev-dependencies]
invoke = "^1.5.0"
python-dotenv = "^0.15.0"
# formating
black = "^20.8b1"
isort = "^5.7.0"
# code smells
pylint = "^2.7.2"
bandit = "^1.7.0"
safety = "^1.10.3"
mypy = "^0.812"
# testing
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
# docs
mkdocs = "^1.1.2"
mkdocstrings = "^0.15.0"
mkdocs-gen-files = "^0.3.1"
plantuml-markdown = "^3.4.2"

[tool.black]
# https://black.readthedocs.io/en/stable/installation_and_usage.html
line-length = 120
exclude = "tests/fixtures"

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/options/
# https://black.readthedocs.io/en/stable/compatible_configs.html
line_length = 120
profile="black"
multi_line_output = 3
include_trailing_comma = true
default_section = "THIRDPARTY"
known_first_party = "python_tooling_enterpy_2021"
lines_after_imports = 2
lines_between_types = 1
force_alphabetical_sort = false

[tool.pytest.ini_options]
# https://docs.pytest.org/en/stable/customize.html
norecursedirs = ".git build dist"
addopts = "--cov --cov-config coverage.ini"

[tool.pylint.messages_control]
# https://black.readthedocs.io/en/stable/compatible_configs.html
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "120"
